+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.0].define(version: 2026_08_18_120000) do
+ActiveRecord::Schema[8.0].define(version: 2026_08_18_140000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "fuzzystrmatch"
   enable_extension "pg_catalog.plpgsql"
@@ -1146,7 +1146,9 @@ ActiveRecord::Schema[8.0].define(version: 2026_08_18_120000) do
     t.bigint "board_group_id"
     t.bigint "board_image_id"
     t.index ["board_group_id"], name: "index_word_events_on_board_group_id"
+    t.index ["board_id"], name: "index_word_events_on_board_id"
     t.index ["board_image_id"], name: "index_word_events_on_board_image_id"
+    t.index ["child_account_id", "timestamp"], name: "index_word_events_on_child_account_id_and_timestamp"
     t.index ["child_account_id"], name: "index_word_events_on_child_account_id"
     t.index ["data"], name: "index_word_events_on_data", using: :gin
     t.index ["image_id"], name: "index_word_events_on_image_id"